+++ app/src/main/java/com/infix/musicappv1/data/model/playlist/PlaylistSongCrossRef.kt
@@ -6,7 +6,7 @@ import androidx.room.Entity
 @Entity("playlist_song_cross_ref", primaryKeys = ["playlist_id", "song_id"])
 data class PlaylistSongCrossRef(
     @ColumnInfo("playlist_id")
-    var playlistId: Int = 0,
+    var playlistId: Long = 0,
     @ColumnInfo("song_id")
     var songId: String = ""
 )
